@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from enum import Enum
 import re
-from typing import Any, Dict, Optional
+from typing import Any, Optional
 
 from pydantic import BaseModel, root_validator, validator
 from pydantic.types import constr
@@ -346,16 +346,16 @@ class MetadataRow(BaseModel):
     mel_type: Optional[MelType]
     mel_ulcer: Optional[bool]
 
-    unstructured: Dict[str, Any]
+    unstructured: dict[str, Any]
 
     # See https://github.com/samuelcolvin/pydantic/issues/2285 for more detail
     @root_validator(pre=True)
-    def build_extra(cls, values: Dict[str, Any]) -> Dict[str, Any]:  # noqa: N805
+    def build_extra(cls, values: dict[str, Any]) -> dict[str, Any]:  # noqa: N805
         all_required_field_names = {
             field.alias for field in cls.__fields__.values() if field.alias != 'unstructured'
         }  # to support alias
 
-        unstructured: Dict[str, Any] = {}
+        unstructured: dict[str, Any] = {}
         for field_name in list(values):
             if field_name not in all_required_field_names:
                 unstructured[field_name] = values.pop(field_name)
