@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import Optional
 
 from django.urls.base import reverse
 
@@ -7,7 +7,7 @@ from isic.ingest.models import Cohort
 from .metadata import *  # noqa
 
 
-def make_breadcrumbs(cohort: Optional[Cohort] = None) -> List:
+def make_breadcrumbs(cohort: Optional[Cohort] = None) -> list:
     ret = [[reverse('ingest-review'), 'Ingest Review']]
 
     if cohort:
